seperated the encryption handler to a seperate class which can be used by other classes; and fixed the test cases
