+++ rpc/wallet.ts
@@ -24,15 +24,6 @@ import { PrivateKeyDBFieldEncryption } from '../src/entity/encryption/PrivateKey
 
 const loadedContracts = LoadedContracts.Instance;
 const metrics = MetricsReporter.Instance;
-const ENV = process.env.ENV || 'LOCAL';
-
-export async function setupWalletEncryptionHandler() {
-    if (ENV === 'DEV' || ENV === 'STAGE' || ENV === 'DEMO' || ENV === 'PROD') {
-        Wallet.setPrivateKeyEncryptionHandler(await AwsPrivateKeyDBFieldEncryption.getInstance());
-    } else {
-        Wallet.setPrivateKeyEncryptionHandler(await PrivateKeyDBFieldEncryption.getInstance());
-    }
-}
 
 @RPCNamespace({ name: 'Wallet' })
 export class RPCWallet {
