+++ src/__tests__/eventsubscriptions.ts
@@ -21,6 +21,7 @@ import { Wallet } from '../entity/Wallet';
 import { Project } from '../entity/Contract';
 import { EventSubscription, EventSubscriberAttrs } from '../entity/EventSubscription';
 import { PrivateKeyDBFieldEncryption } from "../entity/encryption/PrivateKeyDBFieldEncryption";
+import { EncryptorContainer } from '../shipchain/EncryptorContainer';
 
 const request = require('request');
 const utils = require('../local-test-net-utils');
@@ -57,8 +58,8 @@ function AsyncGetJSON(url) {
 export const EventSubscriptionEntityTests = async  function() {
 
     beforeAll(async () => {
-
-        Wallet.setPrivateKeyEncryptionHandler(await PrivateKeyDBFieldEncryption.getInstance());
+        await EncryptorContainer.init();
+        //Wallet.setPrivateKeyEncryptionHandler(await PrivateKeyDBFieldEncryption.getInstance());
     });
 
     it(
