+++ src/shipchain/AwsPrivateKeyDBFieldEncryption.ts
@@ -15,9 +15,9 @@
  */
 
 import { PrivateKeyDBFieldEncryption } from '../entity/encryption/PrivateKeyDBFieldEncryption';
-import { DBFieldEncryption } from '../entity/Wallet';
 import EthCrypto from 'eth-crypto';
 import { getAwsSecret } from './utils';
+import { DBFieldEncryption } from '../entity/encryption/DBFieldEncryption';
 
 const ENV = process.env.ENV || 'LOCAL';
